backend/requirements.txt
# Core Backend Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9  # PostgreSQL adapter
alembic==1.13.1  # Database migrations (optional but recommended)




# Environment Variables
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0  # JWT token handling (added)
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.2

pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0  # Email validation for Pydantic (added)

# Phone Verification (for future Twilio integration)
# twilio==8.10.0  # Uncomment when ready to integrate

# For FCM Push Notifications (add later when implementing)
# firebase-admin==6.3.0

# Development Tools (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv
psycopg2-binary


# fro voice 
librosa== 0.11.0
numpy==2.3.5
scipy==1.17.0


firebase-admin==6.4.0